# Hand rolled parser for SlsVersionType.RELEASE (aka x.y.z)