--- a/changelog/@unreleased/pr-463.v2.yml
+++ b/changelog/@unreleased/pr-463.v2.yml
@@ -1,0 +1,6 @@
+type: improvement
+improvement:
+  description: Parsing the simplest `x.y.z` type of `OrderableSlsVersion` no longer
+    uses regexes. Allocation is reduced from ~600 bytes/op -> ~128bytes/op.
+  links:
+  - https://github.com/palantir/sls-version-java/pull/463