--- a/sls-versions/src/main/java/com/palantir/sls/versions/NonOrderableSlsVersion.java
+++ b/sls-versions/src/main/java/com/palantir/sls/versions/NonOrderableSlsVersion.java
@@ -22,7 +22,6 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import com.palantir.logsafe.UnsafeArg;
 import java.util.Optional;
-import java.util.regex.MatchResult;
 import org.immutables.value.Value;
 
 @Value.Immutable
@@ -49,9 +48,9 @@ public abstract class NonOrderableSlsVersion extends SlsVersion {
 
         return Optional.of(new NonOrderableSlsVersion.Builder()
                 .value(value)
-                .majorVersionNumber(Integer.parseInt(groups.group(1)))
-                .minorVersionNumber(Integer.parseInt(groups.group(2)))
-                .patchVersionNumber(Integer.parseInt(groups.group(3)))
+                .majorVersionNumber(groups.groupAsInt(1))
+                .minorVersionNumber(groups.groupAsInt(2))
+                .patchVersionNumber(groups.groupAsInt(3))
                 .type(SlsVersionType.NON_ORDERABLE)
                 .build());
     }