## Before this PR: 416 bytes / op

From 1 minute JFR, the hottest method in my application is currently `java.util.regex`, and the biggest proportion of these Matcher constructions are coming from the lovely `OrderableSlsVersion#safeValueOf`

![image](https://user-images.githubusercontent.com/3473798/122296063-dde2ca80-cef1-11eb-9eab-6158c2e49401.png)

This seems a bit silly, given that most of the time we're trying to parse three numbers with dots in between them.

```
Benchmark                                                         (versionString)  Mode  Cnt     Score      Error   Units
SlsVersionBenchmark.safeValueOf                                           RELEASE  avgt    3   176.778 ±   30.231   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                            RELEASE  avgt    3  7690.289 ± 1225.982  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                       RELEASE  avgt    3   416.021 ±    0.007    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                   RELEASE  avgt    3  7751.897 ± 1363.284  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm              RELEASE  avgt    3   419.352 ±   10.138    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space               RELEASE  avgt    3     0.035 ±    0.064  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm          RELEASE  avgt    3     0.002 ±    0.004    B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                 RELEASE  avgt    3   184.000             counts
SlsVersionBenchmark.safeValueOf:·gc.time                                  RELEASE  avgt    3   103.000                 ms
```

## After this PR: 104 bytes / op
==COMMIT_MSG==
Deserializing a verison of the form `x.y.z` now allocates less. This should mean GC algorithms have less work to do.
==COMMIT_MSG==

```

Benchmark                                                         (versionString)  Mode  Cnt     Score       Error   Units
SlsVersionBenchmark.safeValueOf                                           RELEASE  avgt    3    55.592 ±   109.978   ns/op
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate                            RELEASE  avgt    3  6158.567 ± 11630.568  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.alloc.rate.norm                       RELEASE  avgt    3   104.006 ±     0.002    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space                   RELEASE  avgt    3  6200.873 ± 12743.698  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Eden_Space.norm              RELEASE  avgt    3   104.656 ±    22.149    B/op
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space               RELEASE  avgt    3     0.019 ±     0.067  MB/sec
SlsVersionBenchmark.safeValueOf:·gc.churn.G1_Survivor_Space.norm          RELEASE  avgt    3    ≈ 10⁻³                B/op
SlsVersionBenchmark.safeValueOf:·gc.count                                 RELEASE  avgt    3   147.000              counts
SlsVersionBenchmark.safeValueOf:·gc.time                                  RELEASE  avgt    3    88.000                  ms
```
## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->