@svc-autorelease 2021-06-21
Released [0.18.0](https://github.com/palantir/sls-version-java/releases/tag/0.18.0)